fix/Fix android pushAccentColor color decoding

+++ android/src/main/kotlin/com/exponea/data/ExponeaConfigurationParser.kt
@@ -64,7 +64,7 @@ class ExponeaConfigurationParser {
                 pushIcon = it.toInt()
             }
             map.getOptional<Double>("pushAccentColor")?.let {
-                pushAccentColor = it.toInt()
+                pushAccentColor = it.toUInt().toInt()
             }
             map.getOptional<String>("pushChannelId")?.let {
                 pushChannelId = it
